Force Tuple recast
Found a small bug on Continuous training, when the model is reloaded, the `model.config.patch_size` is loaded as a list. But the code needs a tuple on training for it to work. 
I fixed it recasting the config prior to training like so `model.config.patch_size = tuple(model.config.patch_size)` 

I here propose to force the patch_size to be a tuple by recasting it for the one check that needs it.