Failed attempt to fix issue regarding output axes order
All image outputs (prob, dist, etc.) and points, distances
have the axes order of the CNN, and not of the input image.
This can lead to unexpected results and is also in contrast to
the behavior of all CARE models in csbdeep.

The code in this commit is more of a reminder that this is broken
and indicates some places that need fixing. Major issue is that
the order of the radial distances has to change when the order of
the image axes change. This seems impossible in general, especially
in 3D.