Resolve integer overflow when running on a large 3D image
In a computation like x+y*nx+z*(nx*ny) if the sub-computation y*nx or
z*(nx*ny) overflows the size of a 32-bit integer, we will get a bogus
result even though the final destination is a 64-bit integer.
The solution is to promote at least one operand to a bigger-sized
integer. Since long can also be 32-bit on 32-bit architectures, we just
use explicit named types here.
This patch also fixes a few cases of comparisons of integers with
different signs.

Resolves: #235