Allow training with sparse label masks
Negative values in label masks will turn off all losses for that pixel during training. This is especially useful for training with sparse annotations, i.e. not all pixels of an image have to be annotated.

This is similar to https://github.com/ksugar/stardist-sparse, but all negative label pixels are ignored (not just -1), and the object classification loss is also disabled (if 'n_classes' is not None).